// src/app/Models/VMMealDetail.ts
export interface VMMealDetail {
    MenuDetailID: Number;
    MealID: Number;
    TimeFrom : string;
    TimeTo : string;
    MealName : string;
    DishTypeID: Number;
    DishType: string;
    DishID: Number;
    DishName : string;
    DishDesc : string;
    MenuDate: Number;
    isActive: boolean;
    CreatedDate: string;
    Day: string;
    Lunch: string;
    Dinner: string;
  }